Move localization types out of the implementation
